# OneDrive/Desktop/AI Projects/SEO/seo-meta-inspector-github-kit/seo-meta-inspector-github-kit/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SeoInspectionInput {
  /** @minLength 8 */
  url: string;
}

export interface ErrorResponse {
  error: string;
}

export type FeedbackInputCategory = typeof FeedbackInputCategory[keyof typeof FeedbackInputCategory];


export const FeedbackInputCategory = {
  problem: 'problem',
  suggestion: 'suggestion',
  question: 'question',
  other: 'other',
} as const;

export interface FeedbackInput {
  category: FeedbackInputCategory;
  /**
     * @minLength 10
     * @maxLength 3000
     */
  message: string;
  /** @maxLength 254 */
  email?: string;
  /** @maxLength 0 */
  website?: string;
}

export interface FeedbackReceipt {
  delivered: boolean;
}

export type SeoTagStatus = typeof SeoTagStatus[keyof typeof SeoTagStatus];


export const SeoTagStatus = {
  pass: 'pass',
  warn: 'warn',
  fail: 'fail',
  info: 'info',
} as const;

export type SeoTagCategory = typeof SeoTagCategory[keyof typeof SeoTagCategory];


export const SeoTagCategory = {
  core: 'core',
  social: 'social',
  technical: 'technical',
  content: 'content',
} as const;

export interface SeoTag {
  key: string;
  label: string;
  /** @nullable */
  value: string | null;
  status: SeoTagStatus;
  note: string;
  category: SeoTagCategory;
  weight: number;
}

export interface SeoCheckGroup {
  key: string;
  label: string;
  passed: number;
  total: number;
  tags: SeoTag[];
}

export interface SeoPreview {
  title: string;
  description: string;
  url: string;
  siteName: string;
  /** @nullable */
  image: string | null;
  /** @nullable */
  imageAlt: string | null;
  twitterCard: string;
  locale: string;
}

export interface SeoInspection {
  url: string;
  fetchedUrl: string;
  hostname: string;
  title: string;
  description: string;
  score: number;
  scoreLabel: string;
  passed: number;
  warnings: number;
  failures: number;
  checkedAt: string;
  responseTimeMs: number;
  groups: SeoCheckGroup[];
  preview: SeoPreview;
}