# OneDrive/Desktop/AI Projects/SEO/seo-meta-inspector-github-kit/seo-meta-inspector-github-kit/lib/api-zod/src/generated/types/feedbackInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FeedbackInputCategory } from './feedbackInputCategory';

export interface FeedbackInput {
  category: FeedbackInputCategory;
  /**
     * @minLength 10
     * @maxLength 3000
     */
  message: string;
  /** @maxLength 254 */
  email?: string;
  /** @maxLength 0 */
  website?: string;
}