# OneDrive/Desktop/AI Projects/SEO/seo-meta-inspector-github-kit/seo-meta-inspector-github-kit/lib/api-zod/src/generated/types/seoCheckGroup.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SeoTag } from './seoTag';

export interface SeoCheckGroup {
  key: string;
  label: string;
  passed: number;
  total: number;
  tags: SeoTag[];
}