# OneDrive/Desktop/AI Projects/SEO/seo-meta-inspector-github-kit/seo-meta-inspector-github-kit/lib/api-zod/src/generated/types/seoInspection.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SeoCheckGroup } from './seoCheckGroup';
import type { SeoPreview } from './seoPreview';

export interface SeoInspection {
  url: string;
  fetchedUrl: string;
  hostname: string;
  title: string;
  description: string;
  score: number;
  scoreLabel: string;
  passed: number;
  warnings: number;
  failures: number;
  checkedAt: string;
  responseTimeMs: number;
  groups: SeoCheckGroup[];
  preview: SeoPreview;
}