# OneDrive/Desktop/AI Projects/SEO/seo-meta-inspector-github-kit/seo-meta-inspector-github-kit/lib/api-zod/src/generated/types/seoPreview.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SeoPreview {
  title: string;
  description: string;
  url: string;
  siteName: string;
  /** @nullable */
  image: string | null;
  /** @nullable */
  imageAlt: string | null;
  twitterCard: string;
  locale: string;
}