# OneDrive/Desktop/AI Projects/SEO/seo-meta-inspector-github-kit/seo-meta-inspector-github-kit/lib/api-zod/src/generated/types/seoTag.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SeoTagCategory } from './seoTagCategory';
import type { SeoTagStatus } from './seoTagStatus';

export interface SeoTag {
  key: string;
  label: string;
  /** @nullable */
  value: string | null;
  status: SeoTagStatus;
  note: string;
  category: SeoTagCategory;
  weight: number;
}